+++ Changes.md
@@ -6,6 +6,7 @@ Features
 
 - TransformQuery : Added a new node to query the transform for a scene location.
 - BoundQuery : Added a new node to query the bound for a scene location.
+- ExistenceQuery : Added a new node to query whether a scene location exists.
 
 Improvements
 ------------